requirements.txt
aniso8601==9.0.1
attrs==20.3.0
base58==2.1.0
brother-ql==0.9.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
idna==2.6
ipfshttpclient==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
packbits==0.6
Pillow==8.0.1
pinatapy-vourhey==0.1.3
pytz==2021.1
pyusb==1.1.1
PyYAML==5.3.1
qrcode==6.1
requests==2.18.4
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.22
varint==1.0.2
Werkzeug==1.0.1
